Fix running the test suite when NO_COLOR is set in calling env
Fix the test suite to run successfully when NO_COLOR is set
in the environment in which pytest is called.  This is done via adding
a new fixture that "sanitizes" the environment for the duration of test
run.  This way pytest respects NO_COLOR while the tested library itself
produces colored output as expected.